Remove HashableObjectStorage
HashableObjectStorage was a package-private base class which was
only used in ReferenceList. This patch just removes this inheritance,
adds all methods to ReferenceList and removes HashableObjectStorage.
More refactoring in ReferenceList will be done in a follow-up patch.
