+++ tests/unit/ReferenceListTest.php
@@ -351,4 +351,31 @@ class ReferenceListTest extends \PHPUnit_Framework_TestCase {
 		$this->assertTrue( $references->equals( unserialize( $serialized ) ) );
 	}
 
+	public function testRemoveDuplicates_noDuplicatesPresent() {
+		$list = new ReferenceList();
+
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 1 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 2 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 3 ) ) ) );
+
+		$list->removeDuplicates();
+
+		$this->assertEquals( 3, count( $list ) );
+	}
+
+	public function testRemoveDuplicates_duplicatesGetRemoved() {
+		$list = new ReferenceList();
+
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 1 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 2 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 3 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 1 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 2 ) ) ) );
+		$list->attach( new Reference( array( new PropertyNoValueSnak( 4 ) ) ) );
+
+		$list->removeDuplicates();
+
+		$this->assertEquals( 4, count( $list ) );
+	}
+
 }
